ncmpidump and the second array size problems proposed by Katie
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@618 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ CDF3/src/utils/ncdump/vardata.h
@@ -16,9 +16,9 @@ extern "C" {
 
 /* Output the data for a single variable, in CDL syntax. */
 extern int vardata ( const struct ncvar*, /* variable */
-		     size_t [], /* variable dimension lengths */
+		     int64_t [], /* variable dimension lengths */
 		     int, /* netcdf id */
-		     int, /* variable id */
+		     int64_t, /* variable id */
 		     const struct fspec* /* formatting specs */
     );
